+++ .gitignore
@@ -1,5 +1,5 @@
-/.env
-/.prod_.env_prod
+.env
+.prod_.env_prod
 from_slack_message_delete.json
 from_slack_message_edit.json
 from_slack_message_post.json
