+++ src/app/Http/Controllers/SlackToSpreadsheetController.php
@@ -10,6 +10,7 @@ use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Response;
 use App\Http\Requests\SlackToSpreadsheetRequest;
 use App\Models\SlackToSpreadsheet;
+use Illuminate\Contracts\Routing\ResponseFactory;
 use Illuminate\Support\Facades\Validator;
 
 class SlackToSpreadsheetController extends Controller
@@ -40,7 +41,7 @@ class SlackToSpreadsheetController extends Controller
             $validated['key_word']
         );
 
-        return response()->json_content('201', 'Resource_Created', 201);
+        return response()->json_content(201, 'Resource_Created', 201);
     }
 
     /**
