+++ src/app/Http/Requests/ApiErrorRequest.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Contracts\Validation\Validator;
+use Illuminate\Http\Exceptions\HttpResponseException;
+
+class ApiErrorRequest extends FormRequest
+{
+    /**
+     * バリデーションチェック失敗時のエラーを投げる
+     *
+     */
+    protected function failedValidation(Validator $validator)
+    {
+        $response = response()->json([
+            'status' => 422,
+            'errors' => $validator->errors(),
+        ], 422);
+
+        throw new HttpResponseException($response);
+    }
+}
